In the beginning there was darkness
If you're reading this it gives me a hope that we share the same
passion about project history and extensive commit messages which are
an essential part of it ;-)

I personally believe that Git is the greatest thing that ever happened
to collaborative software development.

I think it is vital for developers to understand the importance of
maintaining clean Git history of a project in order to prosper in a
longer run.

This is just an initial commit, so nothing really important will be
here, but to honor the fact that you're still reading it here's a
small but funny [collection][1] of other people's initial commit
messages.

Cheers!

[1]:
https://blog.no-panic.at/2014/10/20/funny-initial-git-commit-messages/