:world_map: Design API first
Test assignment says I must submit my API design as a PDF document,
but I'm not sure how one looks like, so for now I'm just going to use
OpenAPI (f.k.a. Swagger) Specification to do this and I'll ask Elliot
from the talent team on the former one later.

I saw somewhere (cannot find where exactly now) you guys using
Swagger (OpenAPI 2.0) to describe your APIs, but for this assignment
I've picked a newer version (3.0) of specification. There are some
[differences][1], but I hope it's not critical for you though.

Although, I was working at @figome for some time and I am more or less
familiar with FinTech domain I still had to made some assumptions in
scope of this coding exercise to design my version of payments
API. Those assumptions are mostly business-specific, so I hope to
learn them while working at @form3tech-oss :bowtie:

The assumptions, for example, are:

* Supported payment schemes (actually, it's not an assumption, but
  examples I took from [Form3 API Postman collection][2])
* Account number codes and formats (I'm sure there're much more of
  those that I have here, but still this is just for demonstration
  purposes)
* I didn't know that UK IBAN format allows letters except for the
  country code (by the way, IBAN format should be hardened to check
  that the first two characters represent a valid country code)

And so on.

An enumeration of currencies that are valid for a payment is, of
course, incomplete. I'm not sure which currencies do you guys allow,
so it's just a list I've made up for the sake of an example :grin:
I've taken some of the hard fiat currencies and add some crypto
flavour to it. Because, well, who knows...

Again, I believe the assumptions I had made are not critical to this
exercise and can be easily adjusted, but, again, it's purely a demo.

[1]: https://blog.readme.io/an-example-filled-guide-to-swagger-3-2/
[2]: https://developer.form3.tech/