:card_file_box: Design database
Design a database for a future application.

Since the [sample][1] I was provided with in the scope of this
exercise doesn't give much information about how data is being
used (for example, whether payers and payees are arbitrary or
registered customers) I decided to put all the attributes into a JSONB
field.

Given more context the database can be designed and normalized much
more efficiently.

[1]: http://mockbin.org/bin/41ca3269-d8c4-4063-9fd5-f306814ff03f