:rocket: Bootstrap app
A couple of notes on dependencies I've picked for this project.

httprouter
----------

In general, for not very complex applications like this one even a
router can be already [too much][1] of dependencies :) Go has rich
standard and HTTP libraries enough to build powerful web
services. Also, it looks like frameworks considered to be an
anti-pattern in Go. On the other hand, packages make us more
productive, so it would be stupid to ignore them and not leveraging
the power of community. Anyways, at the end of the day it's really up
for a team to decide on libraries to use, because often it's [not only
about performance][2] or any other specific reason. Usually, a set of
considerations must be kept in mind in order to pick a library or a
framework. Those, among others, are community support, maintainability
and so on, and performance is just another one among those. Of course,
we are not talking about cases where [performance is business
critical][3] ;) In those scenarios choices are much more obvious.

Anyways, [httprouter][4] looks quite [promising][5] to give it a
try. I figured, this coding exercise a perfect occasion to try it out,
so I took my freedom to give it a shot.

sqlx
----

[sqlx][6] is not really needed, but it just helps to simplify some
database operations and reduces amount of lines of code to be
written.

envconfig
---------

There are plenty of packages to help with configuring your
application. `envconfig` seems to be one of the most popular ones.

Configuring an app in Go is a breeze with default constants and
`os.Getenv()`, but, again, in scope of this assignment I decided to
try it out, too :)

[1]:
https://blog.merovius.de/2017/06/18/how-not-to-use-an-http-router.html
[2]: https://hueniverse.com/performance-at-rest-75bb8fff143
[3]: https://twitter.com/tjholowaychuk/status/1098589765463609345
[4]: https://github.com/julienschmidt/httprouter
[5]: https://github.com/mingrammer/go-web-framework-stars
[6]: https://godoc.org/github.com/jmoiron/sqlx

--
I know, some arguments here come from a Node.js world, but they still
have relevant points.