:sparkles: Add tests
I used to be a unit test nazi which means I always strived for a 100%
test coverage.

Later I've realised that it's always a case for packages and libraries,
but not always necessary for business applications.

This [blog post][1] explains it pretty well :+1:

:warning: These are integration tests! This means they require a running
migrated database to work. Please, refer to 7d04677 if you would like to
use Docker Compose as infrastructure for this service.

[1]: https://kentcdodds.com/blog/write-tests